requirements.txt
requests
beautifulsoup4
ultralytics
opencv-python-headless
psycopg2-binary
python-dotenv
fastapi
uvicorn
jinja2